PA-745: Remove override
This was a fix for the incorrect location of the input labels. However: this issue only occurs because the font is loaded in the browser in a delayed fashion if `disable cache` of the devtools is enabled. In a normal user scenario (e.g. storybook) the label is broken with these overrides, but correct without them.