Revert checkbox commits
This reverts commits b670e4f3 and 3c45680.
I'm reverting on behalf of Daniel Merz as he's not available for the
rest of the week.

Currently, the checkbox doesn't need to be in corporate design as it has
no special logic / styling applied.
We usually want to follow the approach of first devoping components in
the respective webclients and later extract them into the corporate
design once they are stable.

This revert is important to be able to follow semver. Luckily, there was
no release with this component included in the meantime. Otherwise,
removing would require a new major version.