PA-6064 Review: fix order for tasks when sorting on start (asc) in test
Not fixed: the tests uses a calledWith on a callback-spy although the  consumer of the component would expect, that the callback is only called once. Thus calledOnceWith() should be used. But then, the tests would break. To fix this, PA-6292 was created.
This was also the reason, why the tests weren't broken before. The callback was called the first time with the previous order and the second time with the now existing order.