PA-10615 - restructure tooltip logic to get rid of Firefox workaround
Also with the applied workaround the tooltip was shown when the user used "esc" to cancel an open select. During the small delay used in the workaround the tooltip was sometimes displayed for a short amount of time.

The new implementation only wraps a tooltip  around the component if the select is closed.